code cleanup: calibration details binning as int, not double
The binning value can be stored as an int, not a double.

This is backwards compatible with stored calibration values
as the stored values can be read back as integer values
even though they were written out as doubles.

+++ src/scope.h
@@ -232,7 +232,7 @@ public:
     virtual ~Scope();
 
     void SetCalibration(const Calibration& cal) override;
-    void SetCalibrationDetails(const CalibrationDetails& calDetails, double xAngle, double yAngle, double binning);
+    void SetCalibrationDetails(const CalibrationDetails& calDetails, double xAngle, double yAngle, int binning);
     virtual void FlagCalibrationIssue(const CalibrationDetails& calDetails, CalibrationIssueType issue);
     bool IsCalibrated() const override;
     bool BeginCalibration(const PHD_Point& currentLocation) override;
